Add files via upload
KEY SKILLS: (React Native/ Expo/Firebase)
In this app two Bottom tabs are available one is Recent Expense in which  list of last 7 days expense are shown and total amount of last 7 days expense. and other is all expense tab in which all expense are shown all the expense item are clickable which can be update and delete expense. 
on top of right corner  one add {+} icon is available with the help of this user can add new expense .
I use Firebase real time database for storing the expense